refactor: extract handlers from index.ts into modular architecture
Refactor the monolithic index.ts (3,122 lines) into 13 focused handler modules:

Handler modules created:
- blockchain-handlers.ts (254 lines) - Core blockchain tools
- domain-handlers.ts (146 lines) - Domain resolution tools
- transaction-handlers.ts (275 lines) - Transaction & block tools
- token-handlers.ts (138 lines) - Token tools
- multichain-handlers.ts (179 lines) - Multi-chain tools
- contract-handlers.ts (92 lines) - Smart contract tools
- utility-handlers.ts (142 lines) - Utility tools
- endpoint-handlers.ts (236 lines) - Endpoint management tools
- solana-handlers.ts (501 lines) - Solana tools
- cosmos-handlers.ts (746 lines) - Cosmos SDK tools
- sui-handlers.ts (567 lines) - Sui tools
- docs-handlers.ts (152 lines) - Documentation tools
- handlers/index.ts (17 lines) - Central export module

Benefits:
- 95% reduction in main index.ts size (3,122 → 165 lines)
- Clear separation of concerns by blockchain type and functionality
- Easier navigation and maintenance
- Parallel development support
- Isolated testing capabilities
- Reduced merge conflicts

All 70+ tools preserved with identical functionality.
All 175 tests pass successfully.
